Reorganized project as a library
Header files were moved into the include/ directory; example programs like the cat implementation were moved into examples/ directory. CMakeLists was updated to target cxxeffect as a library.
