+++ examples/cat.cpp
@@ -1,6 +1,6 @@
-#include "eff.hpp"
-#include "stream.hpp"
-#include "task.hpp"
+// #include <cxxeffect/eff.hpp>
+#include <cxxeffect/stream.hpp>
+#include <cxxeffect/task.hpp>
 
 int main() {
     eff::stream<eff::task, char> in = eff::io::cin<eff::task>(1024);
